Change for qnx(photon) platform only. It should not affect runtime/buildig other platforms.
Temporary fix until PR:22930 is fixed in libph. The library insist in making any PtRegion widget
sensitive and opaque to key events. Thi shappen if after realizing it, a resource ( like
Pt_ARG_POS or Pt_ARG_DIM ) is set. Such a region will completely block any key events from reaching
any application, if the mouse is in top of it. To test open the url bar, place the mouse in top
of the list and type something. The key events are not reaching anybody.