Modified build_java.pl to create a MANIFEST.MF file with JSS product version information.  This is then added to xpclass(_dbg).jar using jar -tvf in place of zip -R.
Modified all.pl to execute a new test that checks for product package
version information in jss4.jar as well as compare this to the version
of library (libjss4.so or jss4.dll) being used.  This helps in consistency
check for jss jar and associated lib.

Added a new test JSSPackaeTest.java that reads MANIFEST.MF from jss jar
to fetch product version information.