supplimental fix for bug #17230.
we're using the "Rename Address Book" dialog as the "local addressbook properties"
dialog, so let's make that the title.

(it is what you open when you do "Properties" on a local ab)

instead of alerting, allow the user to open it for the PAB or CAB, but
disable the text field.

r=cavin