Fix bug 397036: NPE with FEATURE_ENHANCED_JAVA_ACCESS
What was going on was a java.lang.SecurityException: "Can not make a
java.lang.Class constructor accessible" was getting thrown. So I've
restructured the new code for dealing with getting access to privates to handle
SecurityExceptions and just use accessible members.

I ran with Context.FEATURE_ENHANCED_JAVA_ACCESS set to true over the regression
suite and it turned up a set of cases where a NPE was thrown because a bad
scope object was passed down from ScriptRuntime.newCatchScope. I fixed that and
tightened up the case where a ClassCache isn't found (which was what was
happening with the invalid scope that later resulted in the NPE.